Match items fuzzily instead of literally
The algorithm is very simple and prioritizes by the length of the match
and the match's starting position. It does not factor in starting
letters of words.

`!item ring` yields info for doran's ring
`!item tri` yields info for triforce
`!item tf` is < 3 characters and ignored
`!item t f` yields locket, with triforce getting the same score
`!item dusk` yields info for "Duskblade of Draktharr"
`!item bortk` yields "Blade of the Ruined King" (but not `bork`)